requirements-rag.txt
sentence-transformers
faiss-cpu
numpy
scipy
scikit-learn
pymupdf
beautifulsoup4
requests
pdfminer.six
pillow
python-dotenv
playwright
pyyaml
tqdm
